Refactor consolidate request logic
